Accepted
12	8592 2018-08-02 12:10:26

+++ 206.矩形的个数.cpp
@@ -3,7 +3,7 @@
 using namespace std;
 /* run this program using the console pauser or add your own getch, system("pause") or input loop */
 int main(int argc, char** argv) {
-	int m,n;
+	long long m,n;
 	unsigned long long c;
 	while(cin>>m>>n)
 	{
